Misc fix on kernelconfig + rocmtargetuitls + lowertoukernel.
